Mining Cleanup (#22)
* Modify proof of stake resize

* Update miner.cpp

* Update wallet.cpp

Don't want to continue if nBalance is zero, as we're not staking zerocoins, and don't know what happens if we have a really big number going into SelectStakeCoins

* Minor ENABLE_WALLET cleanup

To prevent compilation errors when building without ENABLE_WALLET

* Abandon CreateNewBlockWithKey() if creation during incorrect phase

CreateNewBlockWithKey() is called from both the Miner and Minter threads, but doesn't abort the block creation until way deep into the creation.  Don't bother wasting the time and resources; and abort earlier.

* Some more bitcoin miner cleanup

* Add BlockValue to vout[2] when splitting a stake

This will prevent the potential for payouts to deduct into a negative vout[2].nValue

* Remove added IncrementExtraNonce

Not sure how that works with Pivx; but IncrementExtraNonce is messing around with vtx[0].vin[0], the coinbase transaction... when there shouldn't be a coinbase transaction for PoS.  This could be a current roadblock.

* Get the right Masternode Payment

Need to calculate the masternode payment for the block height for this new block, not the previous block.  Otherwise we will fail consensus on switchover, as the seesaw won't be balanced correctly.

* Debugging Additions

Additional debugging added (plus a cleanup)