Slow down retries on PoW mine in PoS phase
Rather than shutdown the PoW mining threads on acceptance of the last PoW block, we will just slow down the attempts to keep trying the block; and keep the miner threads running into 6 confirmations of the last PoW block.  This way if somehow the chain is rewound back to proof of work, we're still working.  It's debatable if we want to keep working if we accepted a block that was ultimately rejected; but this is better than a block being rejected and all the miner wallets have already shut down their threads.