Add cs_main TRY_LOCK to makeBalance
`WalletModel::SendCoins` used to call `checkBalanceChanged()` and that was changed to call makeBalance().  We know `IsTrusted()` needs the cs_main lock.

`checkBalanceChanged()` had a `TRY_LOCK` for cs_main in it; which was removed.  That was called from `pollBalanceChanged` as well as `SendCoins`.  Since `pollBalanceChanged()` did a TRY_LOCK, then called into `checkBalanceChanged()`, which also does a TRY_LOCK, it is presumed that a TRY_LOCK in `makeBalance() will be functional if called from `pollBalanceChanged()`, since it's the same flow as the previous code.

Along those same lines; the cs_main lock was taken out on a full scope of the function; which includes the remaining portions of the `CheckBalanceChanged()` routine.  So we will try for the lock still in `CheckBalanceChanged`